+++ internal/domain/error.go
@@ -1,1 +1,7 @@
 package domain
+
+import "errors"
+
+var (
+	NotFoundUser = errors.New("user doesn't exist by mail")
+)
